Style cleanup
